requirements.txt
asgiref==3.9.1
click==8.2.1
colorama==0.4.6
Django==5.2.6
djangorestframework==3.16.1
joblib==1.5.2
nltk==3.9.1
pika==1.3.2
regex==2025.9.1
sqlparse==0.5.3
tqdm==4.67.1
tzdata==2025.2
